requirements.txt
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alabaster==0.7.12
async-generator==1.10
async-timeout==4.0.2
asyncio-dgram==2.1.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.2.0
Babel==2.10.3
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.9
chat-exporter==2.5.1
click==8.0.3
colorama==0.4.4
discord.py==2.3.1
dnspython==2.2.1
docutils==0.19
emoji==2.2.0
frozenlist==1.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
googletrans==3.1.0a0
grapheme==0.6.0
gTTS==2.2.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mcstatus==9.3.0
MinePI==0.5.0
motor==3.0.0
multidict==5.2.0
packaging==21.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pymongo==4.1.1
PyNaCl==1.5.0
pyparsing==3.0.6
pyreadline3==3.4.1
python-dotenv==0.19.2
pytz==2022.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinx-prompt==1.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
typing_extensions==4.0.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
